+++ pkg/ginutil/recover.go
@@ -1,4 +1,4 @@
-package ginutils
+package ginutil
 
 import (
 	"fmt"
